Standardized action function name to simplify CLI functions
