+++ SALib/analyze/delta.py
@@ -145,7 +145,7 @@ def cli_parse(parser):
     return parser
 
 
-def run_analysis(args):
+def cli_action(args):
     problem = read_param_file(args.paramfile)
     Y = np.loadtxt(args.model_output_file,
                    delimiter=args.delimiter, usecols=(args.column,))
@@ -158,4 +158,4 @@ def run_analysis(args):
 
 
 if __name__ == "__main__":
-    common_args.run_cli(cli_parse, run_analysis)
+    common_args.run_cli(cli_parse, cli_action)
