+++ SALib/sample/fast_sampler.py
@@ -88,7 +88,7 @@ def cli_parse(parser):
     return parser
 
 
-def run_sample(args):
+def cli_action(args):
     """Run sampling method
 
     Parameters
@@ -102,4 +102,4 @@ def run_sample(args):
 
 
 if __name__ == "__main__":
-    common_args.run_cli(cli_parse, run_sample)
+    common_args.run_cli(cli_parse, cli_action)
