+++ SALib/sample/saltelli.py
@@ -118,7 +118,7 @@ def cli_parse(parser):
     return parser
 
 
-def run_sample(args):
+def cli_action(args):
     """Run sampling method
 
     Parameters
@@ -134,4 +134,4 @@ def run_sample(args):
 
 
 if __name__ == "__main__":
-    common_args.run_cli(cli_parse, run_sample)
+    common_args.run_cli(cli_parse, cli_action)
